pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mem0-mcp-for-pm"
version = "0.1.0"
description = "MCP Server for mem0 project management integration"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Kunihiro Shimada", email = "kunihiro.shimada@example.com"},
]
maintainers = [
    {name = "Kunihiro Shimada", email = "kunihiro.shimada@example.com"},
]
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.3.0",
    "mem0ai>=0.1.55",
    "python-dotenv>=1.0.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
"Homepage" = "https://github.com/KunihiroS/mem0-mcp-for-pm"
"Bug Tracker" = "https://github.com/KunihiroS/mem0-mcp-for-pm/issues"

[project.scripts]
mem0-mcp-for-pm = "mem0_mcp_wrapper:main"

[tool.hatch.build.targets.wheel]
packages = ["mem0_mcp"]
include = ["mem0_mcp_wrapper.py"]